+++ ansibledoctor/generator/__init__.py
@@ -8,13 +8,15 @@ from ansibledoctor.generator.errors import (
 )
 from ansibledoctor.generator.output_format import OutputFormat
 from ansibledoctor.generator.protocols import DocumentRenderer, TemplateLoader
-from ansibledoctor.generator.renderers import MarkdownRenderer
+from ansibledoctor.generator.renderers import HtmlRenderer, MarkdownRenderer, RstRenderer
 
 __all__ = [
     "OutputFormat",
     "DocumentRenderer",
     "TemplateLoader",
     "MarkdownRenderer",
+    "HtmlRenderer",
+    "RstRenderer",
     "GeneratorError",
     "TemplateError",
     "TemplateNotFoundError",